script with data cleaning outliers
