+++ Pipfile
@@ -7,7 +7,7 @@ verify_ssl = true
 pylint = "*"
 
 [packages]
-e1839a8 = {editable = true,file = "file:///home/yves/Documents/neo/projects/neo/data_science_challenge"}
+e1839a8 = {editable = true,file = "file:///D:\\home\\yves\\Documents\\neo\\projects\\neo\\data_science_project"}
 click = "*"
 coverage = "*"
 awscli = "*"
